requirements.txt
Django==2.1.7
django-cleanup==3.2.0
django-summernote==0.8.11.4
djangorestframework==3.9.3
Pillow==6.0.0
pytz==2019.1
